+++ osu.Game/OsuGame.cs
@@ -67,7 +67,7 @@ namespace osu.Game
 
         private DirectOverlay direct;
 
-        private SocialOverlay social;
+        private DashboardOverlay dashboard;
 
         private BeatmapListingOverlay beatmaplisting;
 
@@ -608,7 +608,7 @@ namespace osu.Game
 
             //overlay elements
             loadComponentSingleFile(direct = new DirectOverlay(), overlayContent.Add, true);
-            loadComponentSingleFile(social = new SocialOverlay(), overlayContent.Add, true);
+            loadComponentSingleFile(dashboard = new DashboardOverlay(), overlayContent.Add, true);
             loadComponentSingleFile(beatmaplisting = new BeatmapListingOverlay(), overlayContent.Add, true);
             loadComponentSingleFile(custommenu = new CustomMenuOverlay(), overlayContent.Add, true);
             var rankingsOverlay = loadComponentSingleFile(new RankingsOverlay(), overlayContent.Add, true);
@@ -669,7 +669,7 @@ namespace osu.Game
             }
 
             // ensure only one of these overlays are open at once.
-            var singleDisplayOverlays = new OverlayContainer[] { chatOverlay, social, direct, changelogOverlay, rankingsOverlay };
+            var singleDisplayOverlays = new OverlayContainer[] { chatOverlay, dashboard, direct, changelogOverlay, rankingsOverlay };
 
             foreach (var overlay in singleDisplayOverlays)
             {
@@ -841,7 +841,7 @@ namespace osu.Game
                     return true;
 
                 case GlobalAction.ToggleSocial:
-                    social.ToggleVisibility();
+                    dashboard.ToggleVisibility();
                     return true;
 
                 case GlobalAction.ResetInputSettings: