+++ osu.Game/Overlays/Dashboard/FriendsBundle.cs
@@ -1,20 +1,24 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Collections.Generic;
 using System.ComponentModel;
+using osu.Game.Online.API.Requests.Responses;
 
-namespace osu.Game.Overlays.Home.Friends
+namespace osu.Game.Overlays.Dashboard.Friends
 {
     public class FriendsBundle
     {
         public FriendsOnlineStatus Status { get; }
 
-        public int Count { get; }
+        public int Count => Users.Count;
 
-        public FriendsBundle(FriendsOnlineStatus status, int count)
+        public List<APIFriend> Users { get; }
+
+        public FriendsBundle(FriendsOnlineStatus status, List<APIFriend> users)
         {
             Status = status;
-            Count = count;
+            Users = users;
         }
     }