+++ osu.Game/Overlays/Rankings/RankingsOverlayHeader.cs
@@ -43,7 +43,7 @@ namespace osu.Game.Overlays.Rankings
             protected override void LoadComplete()
             {
                 base.LoadComplete();//
-                Scope.BindValueChanged(scope => Section = scope.NewValue.GetDescription()??ToString().ToLowerInvariant(), true);
+                Scope.BindValueChanged(scope => Section = scope.NewValue.GetDescription()??ToString(), true);
             }
 
             protected override Drawable CreateIcon() => new ScreenTitleTextureIcon(@"Icons/rankings");