+++ osu.Game/Overlays/SocialOverlay.cs
@@ -123,7 +123,7 @@ namespace osu.Game.Overlays
             {
                 case SocialTab.Friends:
                     var friendRequest = new GetFriendsRequest(); // TODO filter arguments?
-                    friendRequest.Success += users => Users = users.ToArray();
+                    friendRequest.Success += users => Users = users.Select(u => (User)u).ToArray();
                     API.Queue(getUsersRequest = friendRequest);
                     break;